make concrete mattresses support up to 2 decimal places.

+++ src/test/java/uk/co/ogauthority/pwa/model/entity/pwaapplications/form/permanentdeposits/PadPermanentDepositTestUtil.java
@@ -1,5 +1,6 @@
 package uk.co.ogauthority.pwa.model.entity.pwaapplications.form.permanentdeposits;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Arrays;
 import java.util.Set;
@@ -109,7 +110,7 @@ public final class PadPermanentDepositTestUtil {
       Integer entityId,
       String reference,
       PwaApplicationDetail pwaApplicationDetail,
-      int length, int width, int depth,
+      double length, double width, double depth,
       double quantity,
       String contingency,
       LocalDate fromDate,
@@ -124,9 +125,9 @@ public final class PadPermanentDepositTestUtil {
     pd.setMaterialType(MaterialType.CONCRETE_MATTRESSES);
     pd.setQuantity(quantity);
     pd.setContingencyAmount(contingency);
-    pd.setConcreteMattressLength(length);
-    pd.setConcreteMattressWidth(width);
-    pd.setConcreteMattressDepth(depth);
+    pd.setConcreteMattressLength(BigDecimal.valueOf(length));
+    pd.setConcreteMattressWidth(BigDecimal.valueOf(width));
+    pd.setConcreteMattressDepth(BigDecimal.valueOf(depth));
     pd.setFromCoordinates(fromCoordPair);
     pd.setToCoordinates(toCoordPair);
     pd.setToMonth(toDate.getMonthValue());
